+++ src/sip/ua.rs
@@ -92,6 +92,7 @@ impl SipUA {
             outbound_username: cfg.outbound_username.clone(),
             outbound_password: cfg.outbound_password.clone(),
             user_agent: cfg.user_agent.clone(),
+            nat: cfg.nat,
         };
 
         let client = Arc::new(Client::new(client_cfg)?);